# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ledger API",
    "version": "0.1.0"
  },
  "paths": {
    "/accounts": {
      "post": {
        "tags": [
          "accounts"
        ],
        "summary": "Create Account Endpoint",
        "operationId": "create_account_endpoint_accounts_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "List Accounts Endpoint",
        "operationId": "list_accounts_endpoint_accounts_get",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AccountResponse"
                  },
                  "title": "Response List Accounts Endpoint Accounts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/accounts/{account_id}/balance": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "Get Balance Endpoint",
        "operationId": "get_balance_endpoint_accounts__account_id__balance_get",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Account Id"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BalanceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/accounts/{account_id}/transactions": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "Get Account History Endpoint",
        "operationId": "get_account_history_endpoint_accounts__account_id__transactions_get",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Account Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountHistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/settlements/prepare": {
      "post": {
        "tags": [
          "settlements"
        ],
        "summary": "Prepare",
        "operationId": "prepare_settlements_prepare_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettlementPrepare"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettlementBatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/settlements/{settlement_id}/confirm": {
      "post": {
        "tags": [
          "settlements"
        ],
        "summary": "Confirm",
        "operationId": "confirm_settlements__settlement_id__confirm_post",
        "parameters": [
          {
            "name": "settlement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Settlement Id"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettlementConfirmation"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettlementConfirmationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/settlements/{settlement_id}/fail": {
      "post": {
        "tags": [
          "settlements"
        ],
        "summary": "Fail",
        "operationId": "fail_settlements__settlement_id__fail_post",
        "parameters": [
          {
            "name": "settlement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Settlement Id"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettlementFailure"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettlementResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/settlements": {
      "get": {
        "tags": [
          "settlements"
        ],
        "summary": "List All",
        "operationId": "list_all_settlements_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SettlementResponse"
                  },
                  "title": "Response List All Settlements Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountCreate": {
        "properties": {
          "owner_id": {
            "type": "string",
            "title": "Owner Id"
          },
          "account_type": {
            "type": "string",
            "enum": [
              "cash",
              "personal",
              "customer",
              "business",
              "external_bank"
            ],
            "title": "Account Type"
          },
          "currency": {
            "type": "string",
            "maxLength": 3,
            "minLength": 3,
            "pattern": "^[A-Za-z]{3}$",
            "title": "Currency",
            "default": "USD"
          }
        },
        "type": "object",
        "required": [
          "owner_id",
          "account_type"
        ],
        "title": "AccountCreate"
      },
      "AccountHistoryResponse": {
        "properties": {
          "account_id": {
            "type": "string",
            "title": "Account Id"
          },
          "transactions": {
            "items": {
              "$ref": "#/components/schemas/AccountTransaction"
            },
            "type": "array",
            "title": "Transactions"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "account_id",
          "transactions",
          "count"
        ],
        "title": "AccountHistoryResponse"
      },
      "AccountPosting": {
        "properties": {
          "side": {
            "type": "string",
            "title": "Side"
          },
          "amount_minor": {
            "type": "integer",
            "title": "Amount Minor"
          }
        },
        "type": "object",
        "required": [
          "side",
          "amount_minor"
        ],
        "title": "AccountPosting"
      },
      "AccountResponse": {
        "properties": {
          "account_id": {
            "type": "string",
            "title": "Account Id"
          },
          "owner_id": {
            "type": "string",
            "title": "Owner Id"
          },
          "account_type": {
            "type": "string",
            "title": "Account Type"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "account_id",
          "owner_id",
          "account_type",
          "currency",
          "status"
        ],
        "title": "AccountResponse"
      },
      "AccountTransaction": {
        "properties": {
          "transaction_id": {
            "type": "string",
            "title": "Transaction Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "reversal_of_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reversal Of Id"
          },
          "recorded_at": {
            "type": "string",
            "format": "date-time",
            "title": "Recorded At"
          },
          "posting": {
            "$ref": "#/components/schemas/AccountPosting"
          }
        },
        "type": "object",
        "required": [
          "transaction_id",
          "type",
          "state",
          "reversal_of_id",
          "recorded_at",
          "posting"
        ],
        "title": "AccountTransaction"
      },
      "BalanceResponse": {
        "properties": {
          "account_id": {
            "type": "string",
            "title": "Account Id"
          },
          "balance_minor": {
            "type": "integer",
            "title": "Balance Minor"
          },
          "available_balance_minor": {
            "type": "integer",
            "title": "Available Balance Minor"
          }
        },
        "type": "object",
        "required": [
          "account_id",
          "balance_minor",
          "available_balance_minor"
        ],
        "title": "BalanceResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PostingResponse": {
        "properties": {
          "account_id": {
            "type": "string",
            "title": "Account Id"
          },
          "side": {
            "type": "string",
            "title": "Side"
          },
          "amount_minor": {
            "type": "integer",
            "title": "Amount Minor"
          }
        },
        "type": "object",
        "required": [
          "account_id",
          "side",
          "amount_minor"
        ],
        "title": "PostingResponse"
      },
      "SettlementBatchResponse": {
        "properties": {
          "settlements": {
            "items": {
              "$ref": "#/components/schemas/SettlementResponse"
            },
            "type": "array",
            "title": "Settlements"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "settlements",
          "count"
        ],
        "title": "SettlementBatchResponse"
      },
      "SettlementConfirmation": {
        "properties": {
          "external_reference": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "External Reference"
          }
        },
        "type": "object",
        "required": [
          "external_reference"
        ],
        "title": "SettlementConfirmation"
      },
      "SettlementConfirmationResponse": {
        "properties": {
          "settlement": {
            "$ref": "#/components/schemas/SettlementResponse"
          },
          "transaction": {
            "$ref": "#/components/schemas/TransactionResponse"
          }
        },
        "type": "object",
        "required": [
          "settlement",
          "transaction"
        ],
        "title": "SettlementConfirmationResponse"
      },
      "SettlementFailure": {
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "SettlementFailure"
      },
      "SettlementPrepare": {
        "properties": {
          "batch_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Batch Id"
          },
          "external_bank_account_id": {
            "type": "string",
            "title": "External Bank Account Id"
          }
        },
        "type": "object",
        "required": [
          "batch_id",
          "external_bank_account_id"
        ],
        "title": "SettlementPrepare"
      },
      "SettlementResponse": {
        "properties": {
          "settlement_id": {
            "type": "string",
            "title": "Settlement Id"
          },
          "account_id": {
            "type": "string",
            "title": "Account Id"
          },
          "external_bank_account_id": {
            "type": "string",
            "title": "External Bank Account Id"
          },
          "amount_minor": {
            "type": "integer",
            "title": "Amount Minor"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "confirmed",
              "failed"
            ],
            "title": "Status"
          },
          "external_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Reference"
          },
          "correlation_id": {
            "type": "string",
            "title": "Correlation Id"
          },
          "effective_at": {
            "type": "string",
            "format": "date-time",
            "title": "Effective At"
          },
          "confirmed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmed At"
          },
          "transaction_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transaction Id"
          }
        },
        "type": "object",
        "required": [
          "settlement_id",
          "account_id",
          "external_bank_account_id",
          "amount_minor",
          "status",
          "external_reference",
          "correlation_id",
          "effective_at",
          "confirmed_at",
          "transaction_id"
        ],
        "title": "SettlementResponse"
      },
      "TransactionResponse": {
        "properties": {
          "transaction_id": {
            "type": "string",
            "title": "Transaction Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "postings": {
            "items": {
              "$ref": "#/components/schemas/PostingResponse"
            },
            "type": "array",
            "title": "Postings"
          },
          "reversal_of_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reversal Of Id"
          }
        },
        "type": "object",
        "required": [
          "transaction_id",
          "type",
          "state",
          "postings"
        ],
        "title": "TransactionResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}